Crate mail_core

Provides the core mail type Mail for the mail crate. This crate provides the type called Mail as well as ways to create it. It also provides the builder context interface and the Resource type, which is used to represent mail bodies. Especially such which are attachments or embedded images.

Re-exports

pub use context::Context;

pub use context::MaybeEncData;

Modules

compose
This module provides utilities for composing multipart mails.

context
Provides the context needed for building/encoding mails.

default_impl
This module provides an number of default implementations for some of the interfaces.

error
Module containing all custom errors produced by this crate.

mime
Module containing some utilities for MIME usage/creation.

utils
Utilities.

Structs

Data
A type containing some data and metadata for it.

EncData
EncData is like Data but the buffer contains transfer encoded data.

EncodableMail
a mail with all contained futures resolved, so that it can be encoded

IRI
A minimal IRI (International Resource Identifier) implementation which just parses the scheme but no scheme specific part (and neither fragments wrt. those definitions in which fragments are not scheme specific parts).

Mail
A type representing a Mail.

MailFuture
A future resolving to an encodable mail.

Metadata
POD type containing FileMeta, Content-Type and Content-Id

ResourceContainerLoadingFuture

Source
POD containing the IRI which should be used to laod a resource well as an optional file name to use and a description about how the content type should be handled.

Enums

MailBody
A type which either represents a single body, or multiple modies.

Resource
A enum specifying a "resource" for a mail.

TransferEncodingHint
Hint to change how data should be transfer encoded.

UseMediaType
Specifies how the content type should be handled when loading the data.

Traits

ContainedResourcesAccess