Struct mail_core::context::CompositeContext

pub struct CompositeContext<R: ResourceLoaderComponent, O: OffloaderComponent, M: MailIdGenComponent> { /* fields omitted */ }

The CompositeContext is the simplest way to get an Context implementation.

Any custom Context implementations should be realized through the CompositeContext if possible.

This type consists of 3 components it forward all method calls from Context to. This allows the library to have a single Context type but match and mix the parts about resource loading, offloading and id generation in whichever way it fits best.

The composite context will store the components inside of an Arc so that it can be easily shared through an application, it also means non of the components have to implement Clone.

Methods

impl<R, O, M> CompositeContext<R, O, M> where
R: ResourceLoaderComponent,
O: OffloaderComponent,
M: MailIdGenComponent,

pub fn new(resource_loader: R, offloader: O, message_id_gen: M) -> Self

Create a new context from the given components.

pub fn resource_loader(&self) -> &R

Returns a reference to the resource loader component.

pub fn offloader(&self) -> &O

Returns a reference to the offloader component.

pub fn id_gen(&self) -> &M

Returns a reference to the id generation component.