Module mail_core::default_impl

This module provides an number of default implementations for some of the interfaces.

For example it provides a default implementation for the context needed to build/encode a mail.

Modules

simple_context
This module provides a type alias and constructor function for an simple context impl.

Structs

FsResourceLoader
By setting SchemeValidation to Disabled the FsResourceLoader can be used to simple load a resource from a file based on a scheme tail as path independent of the rest, so e.g. it it is used in a Mux which selects a ResourceLoader impl based on a scheme the scheme would not be double validated.

HashedIdGen
a id gen implementation using hash-ing to generate part of it's left hand side

Functions

load_data

simple_cpu_pool
Create a futures_cpupool::CpuPool with default parameters.