Function mail_core::default_impl::simple_cpu_pool

pub fn simple_cpu_pool() -> CpuPool

Create a futures_cpupool::CpuPool with default parameters.

Note that this crate implements OffloaderComponent for CpuPool. (Which is a component of the context used to build/encode mails and should not be confused with header components).