Enum mail_core::Resource

pub enum Resource {
    Source(Source),
    Data(Data),
    EncData(EncData),
}

A enum specifying a "resource" for a mail.

A resource represents any kind of actual data. It can be anything from a html body of a mail over a embedded image to a attached spread sheet.

A resource can be specified in 3 ways:

As a source specifying what to get and how to handle it.

Data (and Metadata) representing a resource.

Data (and Metadata) representing a transfer encoded resource.

Normally auto generated content will be provided as Data, embeddings and attachments will be provided as Source (potentially referring to a file on in a file system) and transfer encoded data can not be provided by the user.

When a mail is converted to a encodable mail any resource will be swapped with a version of it which is transfer encoded, so the only way a consumer of this library normally comes in contact with the third variant is by turning a encodable mail back into normal mail.

Variants

Source(Source)

Provide a source which specify what data to use.

This also allows specifying a media type (if not stored with the data and sniffing is not wanted or to unreliable).

Additionally it allows to specify a "file name" which will force the given name to be used instead of inferring it from the IRI or meta data associated with the IRI in "some way" (like file name fild in a database from which the data will loaded).

Data(Data)

Provide the data used for the mail bodies content.

This for example could be a png image.

EncData(EncData)

Provides a already transfer encoded version of the Data variant.

This can not be created by a consumer of the library and will be created when turning a mail into a transfer encoded mail.

Methods

impl Resource

pub fn load_container<CO>(
container: CO,
ctx: &impl Context
) -> ResourceContainerLoadingFuture<CO> where
CO: ContainedResourcesAccess,

impl Resource

pub fn plain_text(content: impl Into<String>, ctx: &impl Context) -> Resource

Creates a new text Resource with text/plain; charset=utf-8 media type.

The Context is used to generate a ContentId.

pub fn content_id(&self) -> Option<&ContentId>

Return the content id, if there is any.

impl Resource

pub fn create_mail(self) -> Mail

Create a Mail instance representing this Resource.

This is not a complete mail, i.e. it will not contain headers like From or To and in many cases the returned Mail instance will be wrapped into other mail instances adding alternative bodies, embedded resources and attachments.

pub fn create_mail_with_disposition(
self,
disposition_kind: DispositionKind
) -> Mail