Enum mail_core::error::OtherValidationError

pub enum OtherValidationError {
    ContentTypeHeaderGiven,
    ContentTransferEncodingHeaderGiven,
    SingleMultipartMixup,
    InsertSinglepartContentTypeHeader,
    MissingContentTypeHeader,
    NoFrom,
}

Variants

ContentTypeHeaderGiven

Non-multipart mail headers derive the Content-Type header from it's body Resource.

This error is returned if a Content-Type header was given never the less.

ContentTransferEncodingHeaderGiven

Content-Transfer-Encoding headers are always auto-generated and can not be manually set.

SingleMultipartMixup

A non "multipart" media type was given as content type for a multipart mail.

InsertSinglepartContentTypeHeader

Inserting a Conent-Type header into a singlepart body is not allowed.

In single-part bodies the Content-Type header is always auto-generated based on the actual body.

MissingContentTypeHeader

A multipart mail requires a Content-Type header to be given.

NoFrom

A mail (top level, not in multipart) requires a From header to be given.