Struct mail_core::error::ResourceLoadingError

pub struct ResourceLoadingError { /* fields omitted */ }

The loading of an Resource failed.

Methods

impl ResourceLoadingError

pub fn kind(&self) -> ResourceLoadingErrorKind

The kind of error which caused the loading to fail.

pub fn source_iri(&self) -> Option<&IRI>

The source IRI which was used when failing to load the Resource.

pub fn with_source_iri_or_else<F>(self, func: F) -> Self where
F: FnOnce() -> Option<IRI>,

Sets the source IRI if not already set and returns self.