Struct mail_core::EncodableMail

pub struct EncodableMail(_);

a mail with all contained futures resolved, so that it can be encoded

Methods

impl EncodableMail

pub fn encode(&self, encoder: &mut EncodingBuffer) -> Result<(), MailError>

Encode the mail using the given encoding buffer.

After encoding succeeded the buffer should contain a fully encoded mail including all attachments, embedded images alternate bodies etc.

Error

This can fail for a large number of reasons, e.g. some input can not be encoded with the given mail type or some headers/resources breack the mails hard line length limit.

pub fn encode_into_bytes(
&self,
mail_type: MailType
) -> Result<Vec<u8>, MailError>

A wrapper for encode which will create a buffer, enocde the mail and then returns the buffers content.