Module mail_core::utils

Utilities.

Contains everything which doesn't has a better place to be put in.

Structs

Disabled
Struct implementing ConstSwitch with ENABLED = false.

Enabled
Struct implementing ConstSwitch with ENABLED = true.

Traits

ConstSwitch
Trait to allow const bool values in generics.

Functions

now
Returns the current data time.

Type Definitions

SendBoxFuture
Type alias for an boxed future which is Send + 'static.